Remove warning.
[git-p4: depot-paths = "//depot/qt/main/": change = 77483]

+++ tools/assistant/helpdialogimpl.cpp
@@ -975,7 +975,7 @@ void HelpDialog::startSearch()
     terms.clear();
     bool isPhrase = FALSE;
     QString s = "";
-    for ( int i = 0; i < buf.length(); ++i ) {
+    for ( int i = 0; i < (int)buf.length(); ++i ) {
 	if ( buf[i] == '\"' ) {
 	    isPhrase = !isPhrase;
 	    s = s.simplifyWhiteSpace();
